+++ db/migrate/20211122121355_add_taggings_counter_cache_to_tags.acts_as_taggable_on_engine.rb
@@ -8,7 +8,7 @@ else
 end
 AddTaggingsCounterCacheToTags.class_eval do
   def self.up
-    add_column :tags, :taggings_count, :integer, default: 0
+    add_column :tags, :taggings_count, :bigint, default: 0
 
     ActsAsTaggableOn::Tag.reset_column_information
     ActsAsTaggableOn::Tag.find_each do |tag|
